Declutter root package by relocating internal modules
