+++ slide2vec/configs/__init__.py
@@ -1,4 +1,4 @@
-from slide2vec.resources import load_config
+from slide2vec.configs.resources import load_config
 
 
 default_config = load_config("default")
